api/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "richme-api"
version = "0.1.0"
description = "Richme FastAPI backend"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "alembic>=1.14",
  "fastapi>=0.115",
  "passlib[bcrypt]>=1.7",
  "psycopg[binary]>=3.2",
  "pydantic-settings>=2.6",
  "pyjwt>=2.9",
  "sqlalchemy>=2.0",
  "uvicorn[standard]>=0.32",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27",
  "pytest>=8.3",
  "ruff>=0.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
